+++ src/main/java/kitchenpos/application/dto/response/MenuGroupResponse.java
@@ -2,7 +2,7 @@ package kitchenpos.application.dto.response;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroup;
 
 public class MenuGroupResponse {
