+++ src/main/java/kitchenpos/menugroup/service/MenuGroupService.java
@@ -1,10 +1,10 @@
-package kitchenpos.application;
+package kitchenpos.menugroup.service;
 
 import java.util.List;
 import kitchenpos.application.dto.request.MenuGroupRequest;
 import kitchenpos.application.dto.response.MenuGroupResponse;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.repository.MenuGroupRepository;
+import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroupRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
