+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -14,6 +14,7 @@ import kitchenpos.application.dto.response.OrderResponse;
 import kitchenpos.application.dto.response.ProductResponse;
 import kitchenpos.application.dto.response.TableResponse;
 import kitchenpos.menu.service.MenuService;
+import kitchenpos.menugroup.service.MenuGroupService;
 import kitchenpos.product.service.ProductService;
 import kitchenpos.supports.IntegrationTest;
 import kitchenpos.supports.MenuFixture;
@@ -22,6 +23,7 @@ import kitchenpos.supports.OrderFixture;
 import kitchenpos.supports.OrderTableFixture;
 import kitchenpos.supports.ProductFixture;
 import kitchenpos.supports.TableGroupFixture;
+import kitchenpos.tablegroup.service.TableGroupService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
